+++ src/combinator/mod.rs
@@ -162,8 +162,8 @@
 mod branch;
 mod core;
 mod debug;
+mod impls;
 mod multi;
-mod parser;
 mod sequence;
 
 #[cfg(test)]
@@ -172,8 +172,8 @@ mod tests;
 pub use self::branch::*;
 pub use self::core::*;
 pub use self::debug::*;
+pub use self::impls::*;
 pub use self::multi::*;
-pub use self::parser::*;
 pub use self::sequence::*;
 
 #[allow(unused_imports)]
